+++ src/encounters/sszorak/learn2d/Runtime.tsx
@@ -1,0 +1,4 @@
+import EvidenceLearn2DRuntime from '../../../platform/encounters/EvidenceLearn2DRuntime'
+import type { EncounterRuntimeProps } from '../../../platform/encounters'
+import { definition } from '../definition'
+export default function Runtime(props: EncounterRuntimeProps) { return <EvidenceLearn2DRuntime definition={definition} runtime={props} /> }